chore: split DatePicker spec
